+++ src/promise/js/extras/batch.js
@@ -11,5 +11,35 @@ it will be wrapped in a new promise.
 @deprecated @SINCE@
 **/
 Y.batch = function () {
-    return Promise.every(slice.call(arguments));
+    Y.log('batch() was deprecated in YUI @SINCE@. Use Promise.every()', 'warn');
+
+    var funcs     = slice.call(arguments),
+        remaining = funcs.length,
+        i         = 0,
+        length    = funcs.length,
+        results   = [];
+
+    return new Y.Promise(function (fulfill, reject) {
+        var allDone = this;
+
+        function oneDone(index) {
+            return function (value) {
+                results[index] = value;
+
+                remaining--;
+
+                if (!remaining && allDone.getStatus() !== 'rejected') {
+                    fulfill(results);
+                }
+            };
+        }
+
+        if (length < 1) {
+            return fulfill(results);
+        }
+
+        for (; i < length; i++) {
+            Y.when(funcs[i], oneDone(i), reject);
+        }
+    });
 };