+++ src/promise/HISTORY.md
@@ -7,6 +7,7 @@ Promise Change History
 * Split the module into promise-core and promise-extras
 * then() no longer returns an instance of this.constructor
 * Added combinators, factories and extra methods
+* [!] Deprecated Y.batch() in favor of Y.Promise.every()
 * The first function received as parameter of the Promise init function is now
   a reference to resolve() instead of accept()
 * Changing the use of fulfill() to accept() to follow the naming of DOMFuture